extend stdmin.lua & add module base:generation/ores
